+++ deeprankcore/models/operations/parsers/residue.py
