+++ tests/feature/test_pssm.py
@@ -1,7 +1,7 @@
 from pdb2sql import pdb2sql
 import numpy as np
 from deeprankcore.domain.aminoacids import alanine
-from deeprankcore.tools.pssm import parse_pssm
+from deeprankcore.models.operations.parsers.pssm import parse_pssm
 from deeprankcore.models.structure.variant import SingleResidueVariant
 from deeprankcore.features.conservation import add_features
 from deeprankcore.models.graph import build_atomic_graph
