+++ tests/models/test_pssm.py
@@ -1,5 +1,5 @@
 from pdb2sql import pdb2sql
-from deeprankcore.tools.pssm import parse_pssm
+from deeprankcore.models.operations.parsers.pssm import parse_pssm
 from deeprankcore.models.operations.pdb import get_structure
 from deeprankcore.domain.aminoacids import alanine
